feat(bargraph): fractional LED tip + discrete step snapping + T7 UX rework
Bargraph level is now float 0.0-8.0 with fractional tip brightness.
Discrete targets (division, pattern, template) snap to step positions.
Continuous targets get smooth fractional positioning.

LedController: renderBargraph uses fractional tip for caught + uncaught.
ToolLedSettings: remove TAB field cycling, arrows = preset only,
  pot 1 = hue, pot 2 = intensity, preset change resets hue to 0.

+++ src/core/LedController.h
@@ -80,7 +80,7 @@ public:
   void showBatteryGauge(uint8_t percent);
 
   // Pot bargraph with catch visualization
-  void showPotBargraph(uint8_t realLevel, uint8_t potLevel, bool caught);
+  void showPotBargraph(float realLevel, uint8_t potLevel, bool caught);
 
   // Calibration
   void setCalibrationMode(bool active);
@@ -180,7 +180,7 @@ private:
   // Bargraph
   uint16_t _potBarDurationMs;
   bool _showingPotBar;
-  uint8_t _potBarRealLevel;
+  float   _potBarRealLevel;
   uint8_t _potBarPotLevel;
   bool    _potBarCaught;
   unsigned long _potBarStart;
